Debug

+++ concept2vid/extract_concept.py
@@ -309,7 +309,7 @@ def process_input(processor, videos, texts, dtype=torch.float32):
     _, b_t, c, h, w = inputs["pixel_values"].shape
     inputs["pixel_values"] = inputs["pixel_values"].reshape(batch, t, c, h, w).to(dtype)
     logger.info(f'{inputs["pixel_values"].dtype=}')
-    inputs["input_ids"] = torch.Tensor(inputs["input_ids"], dtype=dtype)
+    inputs["input_ids"] = torch.Tensor(inputs["input_ids"]).to(dtype)
     # print(inputs['pixel_values'].shape) # torch.Size([256, 32, 3, 224, 224])
     # print(inputs['input_ids'].shape) # torch.Size([256, 10])
     logger.info(f'{inputs["input_ids"].dtype=}')
